pyproject.toml
[tool.poetry]
name = "cargocode-hack3"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "cargocode_hack3"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-rest-knox = "^4.2.0"
django-filter = "^23.2"
markdown = "^3.4.3"
django-cors-headers = "^4.1.0"
pillow = "^10.0.0"
channels-redis = "^4.1.0"
channels = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
